pyproject.toml
[tool.poetry]
name = "platform"
version = "0.1.0"
description = "To be discussed."
authors = ["Simon Antonius Lauer <simon.lauer@posteo.de>",
           "Georg Vietzke <georg.vietzke@gmail.com>",
           "Mohammed Abu Dunya <mohammed.yahia2020@gmail.com>",
           "Arezou Ramezannezhad <arezou.rgh@gmail.com>",
          ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
six = "^1.16.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
ruff = "^0.6.1"
ipython = "^8.26.0"
django-extensions = "^3.2.3"
psycopg2-binary = "^2.9.9"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
